Code is able to compile and a window is displayed. However, three issues:
1. Text Fragment Shader fails to load properly for an unknown reason
2. Textures aren't being displayed in the window
3. An exception is thrown when key input is being processed. There's issues with dynamic casting an Event into a WidgetEvent, data is frequently lost.

There is also the question of whether the event flow and input reading functionality is maintainable or works properly. The Screen/Scene stack and relationship to the uis might also not be great. For example, instead of a MainMenuScreen or an OptionsScreen pushed and popped onto a stack, maybe have them as some sort of Window or UI object that can hold others and have an object keep track of the the currently active one and the list of root windows?

+++ src/renderer/FontManager.h
@@ -25,7 +25,7 @@ namespace Arboria {
 	inline size_t FontHandle::getHash(const FontHandle& fontHandle) {
 		size_t hash1 = String::getHash(fontHandle.fontName);
 		size_t hash2 = combineHash(hash1, String::getHash(fontHandle.fontStyle));
-		return combineHash(hash2, (fontHandle.fontSize + 119));
+		return combineHash(hash2, (static_cast<size_t>(fontHandle.fontSize) + 119));
 	}
 
 	class FontManager {
